refactor(be): http 통신, 로그인 비즈니스 로직 분리
