+++ server-toss/src/modules/auth/auth.module.ts
@@ -2,10 +2,11 @@ import { Module } from "@nestjs/common";
 import { UserModule } from "src/modules/user/user.module";
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
+import { TossApiClient } from "./toss-api.client";
 
 @Module({
   imports: [UserModule],
   controllers: [AuthController],
-  providers: [AuthService],
+  providers: [AuthService, TossApiClient],
 })
 export class AuthModule {}
